Add an Expression#asStatement overload with JsDoc.
This is helpful for typed declarations like:

/** @private {string} */
my.project.Var;

or

/** @private {number} */
this.num;

This will also be used in a later CL to create a field for the lazy initialization of a parameter's default value.

This also deletes FieldDeclaration, as it was unused.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=241921368