JS runtime support for default parameter values.
Before the parameter type checking code, this generates code to see if the parameter is set. If it's not, it uses the default value instead. If the default value expression is not cheap, it lazily stores the default value in a static property of the template function so it can reuse the same default value instance in future template executions.

For incremental DOM elements, this also declares an instance field for each default parameter. When it calculates the parameter value (as described above) it stores that value in the field so that the JS SoyElement getter functions return the defaulted value.

This also correctly merges default parameter values into data="all" calls.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=242661226