Always allow explicit types on default params and state vars, even if they match the inferred type.
This lets developers decide when to use an explicit type on a default param or state var. This is helpful for cases when the type is hard for humans to infer from the default value (like if it's a complicated expression or a function call) or if being more explicit about the type is beneficial for callers of the template (like if the template is part of a public API).

This was because of suggestion by mrenaud@ on []
GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=243089267