Fix a bug where an unset proto param with a default would fail while trying to dereference the jspbMessageInstance field.
The jspbMessageInstance field is a compatibility mechanism for users who are migrating their templates to use protos from using the toObject form of protos.  For every strictly typed proto parameter we will generate code that looks like

var p = opt_data.p;
p = p.$jspbMessageInstance or p;

to try to recover the original proto from the toObject form.  This code doesn't work for default parameters because for those we allow the parameter to be undefined, which means that p is undefined and this code fails before we can fill in the default.  The fix is just to allow these parameters to be undefined when there is a default expression. So now we will generate code like

var p = opt_data.p;
p = p && (p.$jspbMessageInstance or p)

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=244760989