Attempt to fix template references across af_soy_library boundaries.
Before, the analyzer would hook up template references by finding all callers of a given template. I'm guessing this was broken by []which stopped including the source files of dependencies in the compilation. This meant for a template in a dependency, the analyzer no longer had the source of that dependency so couldn't go through each template in that dependency and find all callers of it to hook up the references.

Now, this loads the Soy header files so the analyzer has details about the dependencies. It finds all template calls, looks up the template definition and infers the VName (based on the file name and template name).

This also adds a delcall test and fixes a TODO about protos.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=247029035