Refactor field management code out of TemplateVariableManager
The addition of @state and default parameters has stretched TemplateVariableManager past its breaking point.  It currently has 3 responsibilities

1. creating field definitions
2. managing local variable lifetimes
3. generating save/restore logic

The issue is that #1 is for the whole class, #2 is for a single method and #3 is for the render() method.

This conflation was fine when the render() method was the only method with any non-trivial logic, but now that we have default parameters and @state initializers we have a desire to generate code for arbitrary soy expressions in the constructor or static initializers.  Right now this is just corrupting some debugging metadata, but future changes will cause more issues.

This requires us to break up these responsibilities.  This is step 1.  next will be to extract a mechanism to create temporary variables that is detatched from save/restore state solution.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=249533688