Small optimization to how let/param gencode captures local variables.
 When a let references a variable defined in the parent we need to 'capture' it since we generate a new class to lazily evaluate the let.  Our previous strategy was to copy every variable into a new field in the closure.  This was simple but was fairly wasteful. for example,

{template .foo}
  {@param p : ?}
  {@param p2 : ?}
  {let $foo: $p + $p2 /}
...
{/template}

That would lead to code that looks like

FooProvider foo = new FooProvider(this.p, this.p2);

...

class FooProvider {
  SoyValueProvider p;
  SoyValueProvider p2;
  FooProvider(SoyValueProvider p, SoyValueProvider p2) {
   this.p = p;
   this.p2 = p2;
  }
...
}

However in the case of parameters (and a few other capturable fields), they are _always_ stored as fields in the parent class.  So instead of copying the values we could just store the parent class directly
class FooProvider {
  FooTemplate $template;
  FooProvider(FooTemplate $template) {
   this.$template = $template;
  }
...
}

and then we can just dereference the template in order to access the parameters.

The one tricky issue is that the fields in the template are all private, so we need to increase the visibility of fields accessed in this way to package private.

this saves about 5K off the compiled size of our integration tests

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=250615932