Remove the localVariableIndex feature from VarDefn
This was added in [] to improve the performance of local variables and to make lazy let/param values possible in tofu, since at the time in tofu variable lookups were based on the string names, which would become ambiguous if a let defined a variable that was the same in an outer scope.

Consider

{let $foo kind="text"}
  {let $a : 1/}
  ...
  {$a}
{/let}
{let $a : 2/}
{$foo}

with the string based lookups when lazily evaluating $foo we would resolve $a to 2 instead of 1.  So instead of using names we switched to using these small unique integers which enabled a more efficient implementation (in both time and memory).

With the benefit of hindsight we should have just used the VarDefn objects as unique keys which is what this CL does.  This does give up on some performance but since only Tofu took advantage of this system it doesn't matter.  Tofu users who are negatively effected by this would be much happier if they migrated to soysauce.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=250888420