Fix how parameter defaults are generated in jbcsrc.
Because the code generating the defaults is adding static constants but was using the render method variable manager, the debug information produced for local variables was incorrect.

The solution is to allocate a separate LocalVariableManager for the <clinit> method and use that in the constant compiler.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=252159441