Don't store trivial lets in fields.
For a certain set of let statements, we can compile them to static final fields.

For example,

{let $len : 16 /}
{for $i in range($len)}
...
{/for}

the $len variable is constant,  so we will stash it in a static final field. e.g.

private static final IntegerData len = IntegerData.forValue(16);

When interacting with the value in the template, we still rely on the variable management code to hand us a local variable and generate 'save/restore' logic for us.  But in the case of values like this the save restore logic is wasteful.  We currently will allocate an extra field to store the value during a detach.

The solution is just to detect this case and avoid the extra work for 'trivial' lets.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=253115142