Remove build nondeterminism by removing use of java.util.Properties.
The class java.util.Properties always includes a comment of the form '#<timestamp' when it is used to serialize the contents of a map to a file.

This timestamp causes build non determinism, which is undesirable.

This change modifies the serialization code to directly write the contents of the map using the format '<key>=<value>\n' instead.

Note that this is slightly different behavior then java.util.Properties.store(). Specifically the store() method will escape spaces and other punctuation (see the documentation on this method for specifics).

GITHUB_BREAKING_CHANGES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=255978832