Change SoyExpression.coerceToBoolean to no longer call itself recursively
The problem with the current approach is that we define a soy expression that 'jumps out of itself'.  This isn't really a problem but due to how it is implemented it means that we cannot use this expression more than once because if we do we will get errors about Labels being used more than once.

In general it isn't exactly clear to me if this is a bug or not or if we should take more steps to prevent this.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=256882816