Fix: Soy VELog use of `xid`
When building Soy using Closure Compiler (OSS), the
`soyutils_velog.js` package can cause issues with its use of
`xid`, which may be a module, which makes it unable to be called
as a function.

This is likely a side effect of the transition to `@idGenerator`-
based `xid` functions.

In any case, installing an `xid` function into the `soyutil_velog`
solves the issue. This changeset does that.