Change JavaScript code constructing JSPB protos to use the 'builder' pattern.
JSPB is a mutable proto API that traditionally had void setter methods.  The gencode was recently changed so that setter methods always 'return this'.  This enables simpler patterns for constructing protos.

This CL attempts to migrate code constructing protos to the new patterns.  It is intended to be a readability improvement and educate users on the new patterns.  If for some reason you prefer to not accept the CL let me know and I will gladly revert.

All formatting changes are courtesy of [] fix (aka clang-format)

[]
Tested:
    [] sample presubmit queue
    []

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=259590472