Automated [] rollback of changelist 260178168.
*** Reason for rollback ***

[] has detected 10 or more targets failed to build at []

TO ROLLFORWARD (without additional approval): Use [] and consider filing a []

To see all broken targets visit []or [] if the former is slow to load.
To prevent noise from flakes, [] double-checked the following target fails to build:
[]
but used to build fine:
[]

Questions? Comments? See the URL: []

*** Original change description ***

Create a new set of passes that executes with template metadata from dependencies but can still modify the template nodes since the metadata has not been created for the current soytree. This will be used to help with [] since we have enough information now in these passes to do inferrence and modify TemplateNodes.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=260202723