Remove custom tokens for misplaced file level declarations
These tokens were added so the parser would eagerly fail when encountering code like

{template .foo}
  {template .bar}
{/template}

however this means that it is not possible write certain kinds of soy expressions (e.g. if a custom soy function has a name that starts with 'element') because the lexer would prefer tokenizing

{elementFoo()}
as
UNEXPECTED_ELEMENT
instead of
PRINT + IDENT

because the lexer always prefers producing longer tokens.

The fix is to remove the token definitions and rely on other parts of the compiler to report different kinds of errors when these mistakes are made.

The error messages are definitely not as good, but im not sure how important that is.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=261157421