Add a new builtin function unknownJsGlobal that will allow access to unresolved 'global' values from the jssrc backend.
Currently, the jssrc backend stands alone in all the soy backends in that it allows unknown globals to be referenced.  This compile to simple namespace references in javascript and might work depending on how things are compiled. This 'feature' traces back into the early days of soy.

To actually make these 'unknown' globals known there are a few approaches:
 * encode the values as proto enums
 * use the soy global system
 * pass the values in as parameters
 * introduce a custom soy plugin to supply the values and dependencies

All these solutions work, but since the 'failure mode' right now is silent there is no encouragement for users to pursue them.  So we will take the same approach we did with v1Expression.  Introduce a builtin function for the deprecated feature and migrate everyone to it.

To that end this CL introduces 'unknownJsGlobal' which:
 * must take a single string literal as a parameter
 * reports an error when used in all backends other than jssrc.

This will allow us to make it an error to use unresolved globals in jssrc just like the other backends and use of unknownJsGlobal can be restricted via the conformance system.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=263015530