Automated [] rollback of changelist 263406371.
*** Reason for rollback ***

Breaks VFE & others...for some reason the polyfill for iterators is not injected.

*** Original change description ***

Rewrite the velog traversal algorithm to account for both velog attributes on elements and <velog> nodes. velog attributes will be used for code size when the element is visible

{velog Foo}
<div></div>
{/velog}

and velog nodes will be used when calls are involved

{velog Foo}
{call .bar /}
{/velog}

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=263437700