Fix: SoyTofu Resolve Template
This patch adds a method to `SoyTofu` to query whether a given
instance holds a template namespace. This change is needed for
compatibility with frameworks downstream (some frameworks check
for a template before rendering).

I'm aware `SoyTofu` is deprecated, but it is still the easiest
path to server-side templates for OSS users. Next up is a similar
implementation for `SoySauce`.

Thank you in advance for considering this change. While `SoyTofu`
is in the rear-view, it would really help our systems work together
with Soy in Java.

Changes:
- [x] Add method to `SoyTofu` interface to query for a template
      namespace
- [x] Add implementation of new method to `BaseTofu`
- [x] Add tests for new method (`hasTemplate` on `SoyTofu`)