Fix: SoySauce Resolve Template
Similar to the change in google/closure-templates#185, this PR adds
a method to `SoySauce` to determine whether a given instance holds a
template. This change is used downstream for frameworks that first
ask a template engine if a template is available, before rendering.

I'm happy to submit these two changes as one. Like the previous PR,
this one comes with unit tests that cover the new functionality.

Changes:
- [x] Add method to `SoySauce` interface, matching the one added in
      `SoyTofu`, called `hasTemplate(String template);`
- [x] Implement new method in `SoySauceImpl` (best way I saw to do
      this is what I did, but open to feedback here of course)
- [x] Add unit test for new functionality