Add Assistant proactive suggestions to jbcsrc/api:module visibility.
We have a LIGHT_UI action in which we render a simple soy template, but within the soy template we make use of a number of soy plugins. We hope to inject and use the @Precompiled SoySauce instance that already has the plugins installed rather than installing the plugins manually on our own SoySauce instance (and having to request visibility on a per plugin basis).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=278419780