Expose methods to go from TypeCheckedVisualElement to Soy VE and vice-versa, in both Java and JS.
The TypeCheckedVisualElement to Soy VE conversion is necessary because we may have TypeCheckedVisualElement embedded in another proto, so we need to convert at runtime in order to end up with a Soy-loggable VE.

The Soy VE to TypeCheckedVisualElement conversion is necessary because we may manually construct such wrapper protos in template code. In the transpiled Soy, these still end up as ve(...) calls, which we can't very well stick on a proto. (The alternative to doing this would be to use TypeCheckedVisualElement everywhere, and converting to the Soy VE type only at the very end when we go to log it, but that seemed like a step backwards - i.e., away from idiomatic Soy - to me).

Also open up the visibility to the Java Soy VE runtime type in preparation for adding a Soy plugin.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=279376644