Drop use of @Precompiled.
It is deprecated and unnecessary. This instead switches the SoyConfigurationModule to use the SoySauce render mode, so that it will create a (un-annotated) SoySauce binding, that can be used for rendering.

This manually configures the one remaining use of Tofu (adds the Soy functions) since this no longer works with the SoyConfigurationModule in SoySauce mode.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=280190302