Index plugins by name in the resolver rather than during SoyFileSet construction.
For SoySourceFunctions the name is stored in an annotation on the class. If the annotation is not present we need to report an error, but currently the location where we extract the name does not have access to an error reporter.

The solution is just to move this work into the PluginResolver.

This does result in a small amount of redundancy in how both the autoescaper and SoySauceImpl deal with print directives, but overall it seems like a simplification.  Plus print directives are deprecated... so in some future this will be simplified once we drop support for user print directives.

GITHUB_BREAKING_CHANGES=It is now an error to name a print directive |map or |record.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=285033420