Automated [] rollback of changelist 285033420.
*** Reason for rollback ***

breaks applications that register the same legacy plugins with guice multiple times

*** Original change description ***

Index plugins by name in the resolver rather than during SoyFileSet construction.

For SoySourceFunctions the name is stored in an annotation on the class. If the annotation is not present we need to report an error, but currently the location where we extract the name does not have access to an error reporter.

The solution is just to move this work into the PluginResolver.

This does result in a small amount of redundancy in how both the autoescaper and SoySauceImpl deal with print directives...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=285079628