Automated [] rollback of changelist 285079628.
*** Reason for rollback ***

rollforward with fix, preserve the behavior where we allow the same class to be regisgtered multiple times.

*** Original change description ***

Automated [] rollback of changelist 285033420.

*** Reason for rollback ***

breaks applications that register the same legacy plugins with guice multiple times

*** Original change description ***

Index plugins by name in the resolver rather than during SoyFileSet construction.

For SoySourceFunctions the name is stored in an annotation on the class. If the annotation is not present we need to report an error, but currently the location where we extract the name does not have access to an erro...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=285292200