Roll forward of changelist 287589885.
NEW: Previous strategy did not account for id resets due to keys.

*** Reason for rollback ***

Breaks this template. This requires resetting the call number when a key node is present

[]

*** Original change description ***

Calculate call keys in a separate pass instead of in GenIncrementalDomCodeVisitor. This is to be used in the jbcsrc pass....

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=288494317