Create ProtoExtensionIdNode to store the fully qualified name parameter the GetExtensionMethod.
Previously, the GlobalNode parameter was converted in the ResolveExpressionTypesPass. If a getExtension method was visited more than once, there would be an error because the first visit would convert the Global into a String, and the second visit would fail because the parameter is no longer a Global. This change allows the visitMethodNode in ResolveExpressionTypesPass to behave the same regardless of the number of times the method is called.

GITHUB_BREAKING_CHANGES=None

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=288970915