Differentiate between use of `FooElement` class and `foo()` function imported from Soy incrementaldom generated code
The code does not compile as written because the `foo(new api.IncrementalDomRenderer(), {})` call refers not to the `foo` variable assigned on the last line, but the `foo()` function generated by the Soy template. Thus we need to rename the variable on the last line as well.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=289765879