Display error range when formatting errors
The benefit of this is that it may be slightly more useful for users and it will help us to debug sourcelocation bugs.

For now we only display a range for source locations on a single line, a followup will attempt to render multiline snippets like typescript.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=290158828