Add a log warning when null is dereferenced in Tofu
if you write something like {if $foo.bar}...  and $foo is null.  This will fail with a runtime error in SoyJs and in SoySauce but in Tofu it will happen to just evaluate as false.  This behavior is difficult to change, but inhibits migrations to SoySauce, so we will at least add a log warning.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=291746764