Automated [] rollback of changelist 292173221.
*** Reason for rollback ***

caused jsdoc formatting issue when a description was wrapped across a line and the next line happened to start with an '@' ¯\_(ツ)_/¯

*** Original change description ***

Preserve newlines in msg.desc attributes

It looks like we have been trimming this since the dark early days of the soy
parser and i don't exactly know why, but from reading the old code, i think it
was essentially an accidental 'feature'.  It should also be fine to preserve
this whitespace in the extracted files.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=292228780