Move some testing utilities to comply with the Java Style Guide: http://[]
This moves AbstractSoyPrintDirectiveTestCase and SharedTestUtils to a "testing" directory in the "java" tree because they are used by multiple test packages. And this moves (and renames) SharedRestrictedTestUtils to the only package its used in, in the "javatests" tree.

This updates the Maven build to exclude the "testing" directory from the main build, and includes it in the testing build.

This also removes an old reference to EscapingConventionsTest since that file was moved in []
GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=294220295