Cache hashcodes from ExprEquivalnce to avoid quadratic behavior in the ResolveExpressionTypesPass
The ResolveExpressionTypesPass uses hashmaps and ExprEquivalence.Wrapper objects to efficiently identify identical subexpressions. However, depending on the structure of the ExprNode tree and the operands in question, this may lead to quadratic runtime behavior.  In particular, and expression like:

  $a and $b and $c ...

Will trigger $a to be hashed N times (where n is the number of 'and' operators), $b will be hashed N-1 times and so on.  By caching the wrapper types and having the wrappers cache their own hashcodes, we can fix this behavior such that we only do a linear amount of work.

This adds some complexity to the API since we now need to decide how long we should cache the hashcodes.  To support this, ExprEquivalnce was rewritten:
  * It is no longer a subtype of Equivalence, though it exports a similar API.  This allows us to manage the lifetime of the Wrapper objects
  * It is no longer a singleton, instead passes that need this will need to manage the lifetime of the object.

This does make using ExprEquivalence a bit more complicated, but for most users this is not an issue.  It should also make it easier to add configuration to ExprEquivalence since different users have slightly different requirements.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=297906908