Parser and AST for the new 'template literal' syntax.
Note that I created a new placeholder type called 'NamedTemplate'. This type stores a reference to the name, and possibly in the future a map of parameters that have already been bound to it, and does no validation. Validation is performed in a separate pass, after the TemplateRegistry is available.

Not in this CL:
 - The new bind() built-in function;
 - Support for {call}-ing template type variables;
 - Support in any of the various Soy runtimes.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=298416142