Fix a small bug in the type checker where we failed to assign types to variant expressions.
This was a regression introduced by the addition of default parameters.  We used to generically handle all exprs in TemplateNodes, but when we added special handling for defaults we forgot to support the other expressions.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=298437154