Fix a small bug in BaseSoyTemplateImpl.asNumber
The intent of this method was to normalize all Number subtypes to either Long
or Double, the actual effect was that it always coerced all Number objects to
Double, due to how java coerces values in ternary statements.

The fix is simple (use an if instead of a ternary).  The difference in behavior
here is almost invisible and few templates could actually depend on the
difference.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=301172097