Add support for comparing records and record fields in ParamTruthSubject
Also add better support for unpacking records in SoyValueUnconverter

This is in preparation for adding better typing for SoyRecords in BaseSoyTemplateImpl.  Right now the main SoyRecord implementation that is used in these places is DictImpl, which also implements SoyMap so records are being handled as maps accidentally.

In preparation for a record implementation that isn't also a map, we will need this support.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=302460493