Refactor null safe access to allow for further improvements.
This creates a new NullSafeAccessNode to represent a null safe access. This structures the AST so that the leftmost null safe access is the parent of then entire access chain. This more accurately reflects the control flow of null safe accesses.

This makes it easier to properly type an access chain with a null safe access as nullable. This also makes it easier to represent access chains with null safe accesses and the forthcoming nonnull assertion operator (so that for an expression like $r?.foo!.bar the control flow exits if $r is null and doesn't execute the nonnull assertion).

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=302892935