Always generate setters for direct params even if there is a conflicting indirect parameter.
We can trust that the compiler will report type errors if necessary.

A case I stumbled across was

{template .a}
  {@param p: string}
  {call .b data="all" /}
{/template}

{template .b}
  {param p: string|uri}
  ...
{/template}

The runtime of soy requires that the parameter be passed as a string so we should just support that in the invocation builders.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=304205073