Automated [] rollback of changelist 305076845.
*** Reason for rollback ***

[]
*** Original change description ***

Improve source locations for data accesses (item, field and method).

The source locations now include the base expression, with an option to get just the source location of the access.

This does the same thing as [] but does it for all DataAccessNodes.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=305141534