Wrap command attribute expressions in ExprRootNodes earlier.
Before this there were two references to these attribute expressions, one was in the CallNode#attributes field and the other where the expression was extracted into its own field, like CallDelegateNode#variantExpr. This meant that if the topmost node of CallDelegateNode#variantExpr was replaced with another node, this wouldn't be reflected in the expression in CallNode#attributes (because it would still have a reference to the replaced node). By ensuring both hold a reference to the same ExprRootNode, these are better kept in sync.

This fixes the issue reported in []which caused the rollback in []. That issue was likely introduced by []and became apparent with []. When NullSafeAccessPass restructed null safe accesses, it would replace the root node of a data access chain. This would work as expected for the reference to the expression from places like CallDelegateNode#variantExpr. But the CallNode#attributes field would still keep a reference to the root of the old expression, which after NullSafeAccess pass was an invalid fragment of an expression. Now that both places are ExprRootNodes, changes to the root node will update both places.

[]

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=305474362