Roll forward of []Improve source locations for data accesses (item, field and method).
The source locations now include the base expression, with an option to get just the source location of the access.

This does the same thing as [] but does it for all DataAccessNodes.

NEW: This only updates golden files, as they had been changed by other CLs.

Automated [] rollback of changelist 305141534.

*** Reason for rollback ***

Roll forward, now that []is submitted, which fixes the issue that caused the rollback.

*** Original change description ***

Automated [] rollback of changelist 305076845.

*** Reason for rollback ***

[]
*** Original change description ***

Improve source locations for data accesses (item, field and method).

The source locations now include the base expression, with an option to get just the source location of the access.

This does the same thing as []

***

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=306226355