Change basic template calls to instantiate templates via the classloader for templates declared as dependencies.
Before this change, all basic template calls were compiled as direct calls to the template's constructor. This implied that a callee template would have to be loaded by the same classloader.

With this change, we limit direct constructor calls to templates that are guaranteed to be loaded by the same classloader as the caller, i.e., templates that appear in the same SoyFileSetNode ancestor and private templates. Otherwhise, we start using Soy's registered classloader to load templates, which opens up the opportunity for more complex classloading strategies.

GITHUB_BREAKING_CHANGES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=308253533