Roll forward of []Roll forward of []Fix VE logging in JBC SRC through ternary and null-coalescing expressions.
This compiles the expressions to SoyValueProviders if either of the sub-expressions can be compiled to SoyValueProviders. This is important because converting a SoyValueProvider with logging statements in it to a SoyValue will drop those logging statements. So values coming into these expressions as SoyValueProviders must stay SoyValueProviders to preserve potential logging statements.

This CL may result in additional VE logs, if those VE logs were previously (incorrectly) dropped by ternary or null-coalescing expressions.

GITHUB_BREAKING_CHANGES=none, unless something was relying on the broken behavior of VE logs being dropped by ternary or null-coalescing expressions.

NEW: Fixed a JBC SRC code generation error.

Automated [] rollback of changelist 308747823.

*** Reason for rollback ***

Roll forward, correctly typing a ternary expression as a SoyValueProvider (in ExpressionToSoyValueProviderCompiler) so downstream code knows to resolve it before using it. This also adds a test demonstrating the bug.

*** Original change description ***

Automated [] rollback of changelist 308628392.

*** Reason for rollback ***

https://paste.[].com/4610809911050240

*** Original change description ***

Roll forward of []Fix VE logging in JBC SRC through ternary and null-coalescing expressions.

This compiles the expressions to SoyValueProviders if either of the sub-expressions can be compiled to SoyValueProviders. This is important because converting a SoyValueProvider with logging statements in it to a SoyValue will...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=309250407