BEGIN_PUBLIC use tslib.d.ts, not tslib_closure.d.ts
The only reason we use the name 'tslib_closure' is for the Closure
.js implementation.  There's no reason to have a separate
tslib_closure.d.ts.

The changes here to tslib.d.ts are copying the changes in
tslib_closure.d.ts to it.  I'm not sure why we diverged from upstream
here, but using the same filename as upstream will make this easier
to track in the future.
END_PUBLIC

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=310200260