Preserve more blank lines.
Left a todo for type nodes, because the blank lines sometimes get moved outside the type in the CommentDecorator. The binding could also be improved for proto inits.

This does cause a slight partial formatting regression where a blank line at the beginning of the parent's rectangle causes us to not really know where to "start" formatting a child (since the parent's source location starts at column 1 from the blank line) -- so, we just format the node starting from wherever it was in the original. I put in a todo for [] to improve the logic here.

GITHUB_BREAKING_CHANGES=NA

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=311347625