Generate JBC SRC code for proto union accesses.
This generates code to check the type of the proto at runtime and call the correct getter method, instead of using reflection.

This also updates proto enums to be more consistently treated as ints in JBC SRC, since the runtime representation of them in JBC SRC is always as ints. This is required to properly box and set the runtime type on the result of a proto union access of an enum field.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=312489006