Reimplement unused variable analysis purely as a tricorder check
Currently this is implemented as a compiler warning, but that imposes several limitations, notably it makes it difficult to add a suggested fix.  The current check also has a number of limitations in that it only examines parameters and ignores let variables and other locals.

The fixer implementation addresses these issues and adds suggested fixes.  The fix logic is a bit complicated since in some cases the syntax requires the use of a name, so this also adds a simple suppression protocol, if the variable name begins with 'unused', then we don't flag it.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=315792851