Make our make the tricorder analyzer configurable.
fixers can be enabled by specifying a list of checks either when creating the aspect or when generating a test.

This blew up on me.  The main reason is that command line flags are handled in 3 places:
1. FindingsGenerator
2. AbstractSoyCompiler
3. FindingAspectGenerator

I decided to rewrite the frontends to treat everything like the way a normal soy compiler frontend works.  And then to reduce some duplication of effort i just registered them in UberSoyCompiler.  We still need a separate java_binary rule to avoid linking all the passes into the normal soy compiler, but this means the aspect main method is now just TricorderAnalyzer.

There is still some awkwardness in the MultiPassTricorderAnalyzer and how it intercepts the file reader and with how AST caching works for normal soy compilation.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=316568794