Fix an error recovery bug.
If we failed parsing an expression in an if-statement we forgot to update the token marking the end of the if tag (the '}').  Then depending on where this occurred we would crash with an error that our tokens were in the wrong order.

This is exacerbated by the fact that we fail to dump the error reporter in the case of an unexpected exception during parsing.  That issue will be addressed by [] which removes our hacky caching strategy.

I also reviewed other locations where we called reportTemplateBodyErrorAndSkipTo to ensure we were doing the right thing and added a CheckReturnValue annotation so the compiler can help.

Finally, i added some test coverage for this, which revealed a few more small bugs in error recovery.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=317346594